Create yolo_server.py
When I use the plain yolov5s model, it only works on side views of a car. 

I trained my own model on overhead camera images which is being used right now.

When I use it on just 1 car in my simulation, this code creates 8 bounding boxes, all different sizes, and only one of them fits perfectly, but its confidence is 0.43. There is one box with 0.85, but its length is too big.